+++ actions/admin/delete-user.ts
@@ -18,6 +18,10 @@ export const deleteUser = async (id: string) => {
       select: { adminPrivileges: true },
     });
 
+    if (!user) {
+      return { error: 'userNotFound' };
+    }
+
     if (session.user.adminLevel !== 'SUPER_ADMIN' && user?.adminPrivileges) {
       return { error: 'cannotDeleteAdmin' };
     }
@@ -25,12 +29,10 @@ export const deleteUser = async (id: string) => {
     await prisma.$transaction(async (tx) => {
       await tx.review.updateMany({
         where: { userId: id },
-
         data: {
-          userId: 'deleted-user',
+          userId: '6835d5809150feb71e83d922',
         },
       });
-
       const cart = await tx.cart.findUnique({
         where: { userId: id },
         select: { id: true },
@@ -59,24 +61,79 @@ export const deleteUser = async (id: string) => {
       });
 
       if (store) {
-        await tx.product.deleteMany({
+        const storeProducts = await tx.product.findMany({
           where: { storeId: store.id },
+          select: { id: true },
         });
 
-        await tx.store.delete({
-          where: { id: store.id },
+        await tx.cartItem.deleteMany({
+          where: {
+            productId: { in: storeProducts.map((p) => p.id) },
+          },
         });
+
+        await tx.favoriteItem.deleteMany({
+          where: {
+            productId: { in: storeProducts.map((p) => p.id) },
+          },
+        });
+
+        const productsWithOrders = await tx.orderItem.findMany({
+          where: {
+            productId: { in: storeProducts.map((p) => p.id) },
+          },
+          select: { productId: true },
+        });
+
+        const productsWithOrdersSet = new Set(
+          productsWithOrders.map((p) => p.productId),
+        );
+
+        const productsWithoutOrders = storeProducts.filter(
+          (p) => !productsWithOrdersSet.has(p.id),
+        );
+
+        if (productsWithOrders.length > 0) {
+          await tx.product.updateMany({
+            where: {
+              id: { in: productsWithOrders.map((p) => p.productId) },
+            },
+            data: {
+              isActive: false,
+              quantity: 0,
+              deleted: true,
+              storeId: null,
+            },
+          });
+        }
+
+        if (productsWithoutOrders.length > 0) {
+          await tx.product.deleteMany({
+            where: {
+              id: { in: productsWithoutOrders.map((p) => p.id) },
+            },
+          });
+        }
+
+        if (productsWithOrders.length === 0) {
+          await tx.store.delete({
+            where: { id: id },
+          });
+        } else {
+          await tx.store.update({
+            where: { userId: id },
+            data: {
+              active: false,
+            },
+          });
+        }
       }
 
-      await tx.order.updateMany({
-        where: { userId: id },
-        data: {
-          userId: 'deleted-user',
-        },
-      });
-
-      await tx.user.delete({
+      await tx.user.update({
         where: { id },
+        data: {
+          deleted: true,
+        },
       });
     });
 
@@ -86,8 +143,8 @@ export const deleteUser = async (id: string) => {
   } catch (error) {
     if (error instanceof PrismaClientKnownRequestError) {
       if (error.code === 'P2025') return { error: 'userNotFound' };
-      return { error: 'validationError' };
     }
+    console.log(error);
     return { error: 'validationError' };
   }
 };
